% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinanceHistory.R
\name{BinanceHistory}
\alias{BinanceHistory}
\title{BinanceHistory Class}
\description{
Get history data from Binance.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path_to_dump}}{Local path to save data to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BinanceHistory-new}{\code{BinanceHistory$new()}}
\item \href{#method-BinanceHistory-dump_binance}{\code{BinanceHistory$dump_binance()}}
\item \href{#method-BinanceHistory-clone}{\code{BinanceHistory$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinanceHistory-new"></a>}}
\if{latex}{\out{\hypertarget{method-BinanceHistory-new}{}}}
\subsection{Method \code{new()}}{
Create a new BinanceHistory object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinanceHistory$new(path_to_dump)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_to_dump}}{Local path to save data to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `BinanceHistory` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinanceHistory-dump_binance"></a>}}
\if{latex}{\out{\hypertarget{method-BinanceHistory-dump_binance}{}}}
\subsection{Method \code{dump_binance()}}{
dump historic data from binance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinanceHistory$dump_binance(asset, data_type, frequency)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{asset}}{binance asset. Can be cm, um, option or spot}

\item{\code{data_type}}{binance data type.}

\item{\code{frequency}}{time frequency.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BinanceHistory-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BinanceHistory-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinanceHistory$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
