% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CME.R
\name{CME}
\alias{CME}
\title{CME Class}
\description{
Get data data from CME data API.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_key}}{API KEY for CME}

\item{\code{client_id}}{Client ID for CME}

\item{\code{client_secret}}{Client secret ID for CME}

\item{\code{token}}{Token we use to make requests}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CME-new}{\code{CME$new()}}
\item \href{#method-CME-get_token}{\code{CME$get_token()}}
\item \href{#method-CME-retoken}{\code{CME$retoken()}}
\item \href{#method-CME-fedwatch}{\code{CME$fedwatch()}}
\item \href{#method-CME-get_meetings_history}{\code{CME$get_meetings_history()}}
\item \href{#method-CME-clone}{\code{CME$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CME-new"></a>}}
\if{latex}{\out{\hypertarget{method-CME-new}{}}}
\subsection{Method \code{new()}}{
Create a new CME object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CME$new(api_key = NULL, client_id = NULL, client_secret = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{API KEY for CME cloud data.}

\item{\code{client_id}}{Client ID for CME cloud data.}

\item{\code{client_secret}}{Client Secret for CME cloud data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `CME` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CME-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-CME-get_token}{}}}
\subsection{Method \code{get_token()}}{
Get CME token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CME$get_token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
CME token.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CME-retoken"></a>}}
\if{latex}{\out{\hypertarget{method-CME-retoken}{}}}
\subsection{Method \code{retoken()}}{
Refresh the authentication token if it's expired or missing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CME$retoken()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CME-fedwatch"></a>}}
\if{latex}{\out{\hypertarget{method-CME-fedwatch}{}}}
\subsection{Method \code{fedwatch()}}{
Get the current FedWatch forecasts with full API parameter support.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CME$fedwatch(meeting_date = NULL, reporting_date = NULL, limit = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meeting_date}}{Optional. Filter forecasts by specific meeting date in 'yyyy-mm-dd' format.}

\item{\code{reporting_date}}{Optional. Filter forecasts by specific reporting date in 'yyyy-mm-dd' format.}

\item{\code{limit}}{Optional. Limits the number of results returned. Default is 10.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the FedWatch forecast data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CME-get_meetings_history"></a>}}
\if{latex}{\out{\hypertarget{method-CME-get_meetings_history}{}}}
\subsection{Method \code{get_meetings_history()}}{
Get historical meetings data from the CME FedWatch API.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CME$get_meetings_history(limit = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{An integer specifying the number of records to retrieve. Default is 10.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the meetings history data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CME-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CME-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CME$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
