% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Data}
\alias{Data}
\title{Data Class}
\description{
Get or update all data from the package.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fmp}}{FMP object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Data$new()}}
\item \href{#method-get_all_data}{\code{Data$get_all_data()}}
\item \href{#method-clone}{\code{Data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Data object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$new(fmp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fmp}}{FMP object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Data` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_all_data}{}}}
\subsection{Method \code{get_all_data()}}{
Scrap (download) all data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$get_all_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Earning announcements data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
