% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAbstract.R
\name{DataAbstract}
\alias{DataAbstract}
\title{DataAbstract Class}
\description{
Abstract class for all other data classes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{context_with_config}}{AWS S3 Tiledb config}

\item{\code{fredr_apikey}}{FRED api key.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataAbstract-new}{\code{DataAbstract$new()}}
\item \href{#method-DataAbstract-clone}{\code{DataAbstract$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAbstract-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataAbstract-new}{}}}
\subsection{Method \code{new()}}{
Create a new DataAbstract object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAbstract$new(context_with_config = NULL, fredr_apikey = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context_with_config}}{AWS S3 Tiledb config}

\item{\code{fredr_apikey}}{FRED api key}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DataAbstract` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataAbstract-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataAbstract-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAbstract$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
