% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dukascopy.R
\name{Dukascopy}
\alias{Dukascopy}
\title{Dukascopy Class}
\description{
Get data data from Dukascopy.
}
\references{
\url{https://www.interactivebrokers.com/api/doc.html}
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{Dukascopy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Dukascopy-new}{\code{Dukascopy$new()}}
\item \href{#method-Dukascopy-get_symbols}{\code{Dukascopy$get_symbols()}}
\item \href{#method-Dukascopy-download_raw}{\code{Dukascopy$download_raw()}}
\item \href{#method-Dukascopy-raw_to_dt}{\code{Dukascopy$raw_to_dt()}}
\item \href{#method-Dukascopy-dukascopy_node_get}{\code{Dukascopy$dukascopy_node_get()}}
\item \href{#method-Dukascopy-clone}{\code{Dukascopy$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dukascopy-new"></a>}}
\if{latex}{\out{\hypertarget{method-Dukascopy-new}{}}}
\subsection{Method \code{new()}}{
Create a new Dukascopy object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dukascopy$new(azure_storage_endpoint = NULL, context_with_config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{azure_storage_endpoint}}{Azure storate endpont}

\item{\code{context_with_config}}{AWS S3 Tiledb config}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Dukascopy` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dukascopy-get_symbols"></a>}}
\if{latex}{\out{\hypertarget{method-Dukascopy-get_symbols}{}}}
\subsection{Method \code{get_symbols()}}{
Get all symbols from Dukascopy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dukascopy$get_symbols()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of symbols.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dukascopy-download_raw"></a>}}
\if{latex}{\out{\hypertarget{method-Dukascopy-download_raw}{}}}
\subsection{Method \code{download_raw()}}{
IB GET object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dukascopy$download_raw(
  save_path = "D:/dukascopy",
  symbols = "SPYUSUSD",
  start_date = as.Date("2017-01-01"),
  end_date = Sys.Date() - 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_path}}{Saving directory.}

\item{\code{symbols}}{Sybmols to scrap.}

\item{\code{start_date}}{Start date.}

\item{\code{end_date}}{End date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GET response.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dukascopy-raw_to_dt"></a>}}
\if{latex}{\out{\hypertarget{method-Dukascopy-raw_to_dt}{}}}
\subsection{Method \code{raw_to_dt()}}{
Parse Dukascopy data and convert to data.table format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dukascopy$raw_to_dt(files)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{files}}{Bin files (lzma files) downlaoded from Dukascopy and saved
localy.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data.table with ask and bid prices and volumes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dukascopy-dukascopy_node_get"></a>}}
\if{latex}{\out{\hypertarget{method-Dukascopy-dukascopy_node_get}{}}}
\subsection{Method \code{dukascopy_node_get()}}{
Get Dukascopy data using https://github.com/Leo4815162342/dukascopy-node
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dukascopy$dukascopy_node_get(
  dir_location,
  instrument,
  from,
  to = "now",
  timeframe = "tick",
  volumes = "true",
  format = "csv",
  batch_size = 5,
  batch_pause = 2500,
  cache = "true"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir_location}}{Location where files are saved.}

\item{\code{instrument}}{Instrument.}

\item{\code{from}}{Date from.}

\item{\code{to}}{Date to.}

\item{\code{timeframe}}{Timeframe aggregation (tick, s1, m1, m5, m15, m30, h1, h4, d1, mn1).}

\item{\code{volumes}}{Include volumes.}

\item{\code{format}}{Output format (csv, json, array).}

\item{\code{batch_size}}{Batch size.}

\item{\code{batch_pause}}{Batch pause.}

\item{\code{cache}}{Use cache.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL or error.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dukascopy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Dukascopy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dukascopy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
