% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eodhs.R
\name{EODHD}
\alias{EODHD}
\title{EODHD Class}
\description{
Get data data from EODHD.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{EODHD}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_key}}{API KEY for EODHD.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EODHD-new}{\code{EODHD$new()}}
\item \href{#method-EODHD-get_intraday_bulk}{\code{EODHD$get_intraday_bulk()}}
\item \href{#method-EODHD-clone}{\code{EODHD$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EODHD-new"></a>}}
\if{latex}{\out{\hypertarget{method-EODHD-new}{}}}
\subsection{Method \code{new()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EODHD$new(api_key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{API KEY for FMP cloud data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FMP` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EODHD-get_intraday_bulk"></a>}}
\if{latex}{\out{\hypertarget{method-EODHD-get_intraday_bulk}{}}}
\subsection{Method \code{get_intraday_bulk()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EODHD$get_intraday_bulk(symbols, uri, workers = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Symbol of the stock.}

\item{\code{uri}}{to save data to.}

\item{\code{workers}}{Number of workers.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Earning announcements data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EODHD-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EODHD-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EODHD$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
