% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eurex.R
\name{Eurex}
\alias{Eurex}
\title{Eurex Class}
\description{
Get data data from Eurex.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{delay}}{Delay between requests.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Eurex-new}{\code{Eurex$new()}}
\item \href{#method-Eurex-get_instruments}{\code{Eurex$get_instruments()}}
\item \href{#method-Eurex-get_markets}{\code{Eurex$get_markets()}}
\item \href{#method-Eurex-get_ticks}{\code{Eurex$get_ticks()}}
\item \href{#method-Eurex-clone}{\code{Eurex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Eurex-new"></a>}}
\if{latex}{\out{\hypertarget{method-Eurex-new}{}}}
\subsection{Method \code{new()}}{
Create a new Eurex object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Eurex$new(delay = 0.8)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delay}}{Delay between requests.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MacroData` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Eurex-get_instruments"></a>}}
\if{latex}{\out{\hypertarget{method-Eurex-get_instruments}{}}}
\subsection{Method \code{get_instruments()}}{
Get instruments from Eurex.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Eurex$get_instruments()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get instruments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Eurex-get_markets"></a>}}
\if{latex}{\out{\hypertarget{method-Eurex-get_markets}{}}}
\subsection{Method \code{get_markets()}}{
Get markets from Eurex.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Eurex$get_markets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Get markets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Eurex-get_ticks"></a>}}
\if{latex}{\out{\hypertarget{method-Eurex-get_ticks}{}}}
\subsection{Method \code{get_ticks()}}{
Get daily data for future symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Eurex$get_ticks(id, start_date = start_date, end_date = end_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Start date}

\item{\code{start_date}}{Start date}

\item{\code{end_date}}{End date}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Get investing com ea data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Eurex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Eurex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Eurex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
