% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{findata-package}
\alias{findata}
\alias{findata-package}
\title{findata: Collect data fro trading and financial ML}
\description{
Collect data for trading and financial ML. Data sources are FMP Cloud, Investing.com, Finam.ru, Dukascopy, FRED.
}
