% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstRate.R
\name{FirstRate}
\alias{FirstRate}
\title{FirstRate Class}
\description{
Get data data from FirstRate.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{FirstRate}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{userid}}{FirstRate user ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FirstRate-new}{\code{FirstRate$new()}}
\item \href{#method-FirstRate-get_historical_data}{\code{FirstRate$get_historical_data()}}
\item \href{#method-FirstRate-firstrate_to_tledb}{\code{FirstRate$firstrate_to_tledb()}}
\item \href{#method-FirstRate-clone}{\code{FirstRate$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirstRate-new"></a>}}
\if{latex}{\out{\hypertarget{method-FirstRate-new}{}}}
\subsection{Method \code{new()}}{
Create a new FirstRate object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirstRate$new(
  userid = NULL,
  azure_storage_endpoint = NULL,
  context_with_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{FirstRate user ID.}

\item{\code{azure_storage_endpoint}}{Azure storate endpont}

\item{\code{context_with_config}}{AWS S3 Tiledb config}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FirstRate` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirstRate-get_historical_data"></a>}}
\if{latex}{\out{\hypertarget{method-FirstRate-get_historical_data}{}}}
\subsection{Method \code{get_historical_data()}}{
Download stock data from FirstRate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirstRate$get_historical_data(
  save_dest,
  type = c("stock"),
  period = c("full", "month", "week", "day"),
  ticker_range = toupper(letters),
  timeframe = c("1min", "5min", "30min", "1hour", "1day"),
  adjustment = c("adj_split", "adj_splitdiv", "UNADJUSTED")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_dest}}{Save path}

\item{\code{type}}{Security type}

\item{\code{period}}{Period.}

\item{\code{ticker_range}}{First letter of symbols.}

\item{\code{timeframe}}{Time frequency.}

\item{\code{adjustment}}{Adjustment type.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirstRate-firstrate_to_tledb"></a>}}
\if{latex}{\out{\hypertarget{method-FirstRate-firstrate_to_tledb}{}}}
\subsection{Method \code{firstrate_to_tledb()}}{
Save data to TileDB Uri.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirstRate$firstrate_to_tledb(
  zip_files,
  save_uri = "s3://equity-usa-hour-firstrate-adjusted"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zip_files}}{Zip files downloaded from FirstRate site.}

\item{\code{save_uri}}{TileDB URI to save data to.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FirstRate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FirstRate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FirstRate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
