% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{getCovar}
\alias{getCovar}
\title{easily get financial asset covariance}
\usage{
getCovar(tickers, start, end)
}
\arguments{
\item{vector}{string formatted stock tickers}

\item{string}{data collection start date}

\item{string}{data collection end date}
}
\value{
matrix with annualized variance/covariance of all assets
}
\description{
easily get financial asset covariance
}
\examples{
getCovar(c("FB","GS"),"2014-12-01","2015-01-01")
}

