% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{getReturns}
\alias{getReturns}
\title{easily get stock return data}
\usage{
getReturns(tickers, look_back, start, end, rolling = FALSE)
}
\arguments{
\item{vector}{string formatted stock tickers}

\item{number}{return period (1->daily returns, 21->monthly, 252->annual)}

\item{string}{data collection start date}

\item{string}{data collection end date}

\item{boolean}{get rolling returns? defaults FALSE}
}
\value{
list of matrices with return data
}
\description{
easily get stock return data
}
\examples{
getReturns(c("FB","GS"),21,"2012-01-01","2015-01-01")
}

