% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBREST.R
\name{IBREST}
\alias{IBREST}
\title{IBREST Class}
\description{
Get data data from IB Client Portal API.
}
\references{
\url{https://www.interactivebrokers.com/api/doc.html}

\url{https://www.interactivebrokers.com/api/doc.html}

\url{https://www.interactivebrokers.com/api/doc.html#tag/Market-Data/paths/~1iserver~1marketdata~1history/get}

\url{https://www.interactivebrokers.com/api/doc.html#tag/Order/paths/~1iserver~1account~1\%7Bacc€ountId\%7D~1orders/post}
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{IBREST}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IBREST-new}{\code{IBREST$new()}}
\item \href{#method-IBREST-ib_get}{\code{IBREST$ib_get()}}
\item \href{#method-IBREST-ib_post}{\code{IBREST$ib_post()}}
\item \href{#method-IBREST-get_api_spec}{\code{IBREST$get_api_spec()}}
\item \href{#method-IBREST-get_unadjusted_market}{\code{IBREST$get_unadjusted_market()}}
\item \href{#method-IBREST-get_position}{\code{IBREST$get_position()}}
\item \href{#method-IBREST-buy_and_confirm}{\code{IBREST$buy_and_confirm()}}
\item \href{#method-IBREST-cancel_order}{\code{IBREST$cancel_order()}}
\item \href{#method-IBREST-clone}{\code{IBREST$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST-new"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST-new}{}}}
\subsection{Method \code{new()}}{
Create a new IBREST object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST$new(azure_storage_endpoint = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{azure_storage_endpoint}}{Azure storage endpoint}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `IBREST` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST-ib_get"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST-ib_get}{}}}
\subsection{Method \code{ib_get()}}{
IB GET object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST$ib_get(url = "https://localhost:5000/v1/api/sso/validate", query = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{Url of GET endpoint.}

\item{\code{query}}{Query of GET request.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GET response.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST-ib_post"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST-ib_post}{}}}
\subsection{Method \code{ib_post()}}{
IB POST object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST$ib_post(url = "https://localhost:5000/v1/api/tickle", body = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{Url of POST endpoint.}

\item{\code{body}}{Body of POST request.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GET response.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST-get_api_spec"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST-get_api_spec}{}}}
\subsection{Method \code{get_api_spec()}}{
Get IB REST API description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST$get_api_spec()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON object with description.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST-get_unadjusted_market"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST-get_unadjusted_market}{}}}
\subsection{Method \code{get_unadjusted_market()}}{
Get unadjusted market data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST$get_unadjusted_market(
  conid,
  exchange = NULL,
  period = "5d",
  bar = "1h",
  outsideRth = TRUE,
  keep_nytime_10_16 = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conid}}{Contract ID.}

\item{\code{exchange}}{Exchange.}

\item{\code{period}}{Period.}

\item{\code{bar}}{Bar.}

\item{\code{outsideRth}}{Data outside trading hours.}

\item{\code{keep_nytime_10_16}}{If TRUE, timezone is changed to NY time and only
trading hours kept.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with unadjusted market data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST-get_position"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST-get_position}{}}}
\subsection{Method \code{get_position()}}{
Get portfolio positions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST$get_position(account_id, con_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_id}}{Account ID.}

\item{\code{con_id}}{Contract id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list object with info on positions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST-buy_and_confirm"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST-buy_and_confirm}{}}}
\subsection{Method \code{buy_and_confirm()}}{
Place order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST$buy_and_confirm(account_id, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_id}}{Account ID.}

\item{\code{...}}{Body of POST request which place order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list object with info on positions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST-cancel_order"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST-cancel_order}{}}}
\subsection{Method \code{cancel_order()}}{
Cancel order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST$cancel_order(account_id, order_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_id}}{Account ID.}

\item{\code{order_id}}{Order id from Place orders endpoint result.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list object with info on positions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
