% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBREST2.R
\name{IBREST2}
\alias{IBREST2}
\title{IBREST Class}
\description{
Get data data from IB Client Portal API.
}
\references{
\url{https://www.interactivebrokers.com/api/doc.html}

\url{https://www.interactivebrokers.com/api/doc.html}

\url{https://www.interactivebrokers.com/api/doc.html#tag/Market-Data/paths/~1iserver~1marketdata~1history/get}

\url{https://www.interactivebrokers.com/api/doc.html#tag/Order/paths/~1iserver~1account~1\%7Bacc€ountId\%7D~1orders/post}
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{IBREST2}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{domain}}{Domain, by default localhost}

\item{\code{port}}{Port, by default 5000}

\item{\code{baseurl}}{Base url}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IBREST2-new}{\code{IBREST2$new()}}
\item \href{#method-IBREST2-ib_get}{\code{IBREST2$ib_get()}}
\item \href{#method-IBREST2-ib_post}{\code{IBREST2$ib_post()}}
\item \href{#method-IBREST2-get_api_spec}{\code{IBREST2$get_api_spec()}}
\item \href{#method-IBREST2-get_unadjusted_market}{\code{IBREST2$get_unadjusted_market()}}
\item \href{#method-IBREST2-get_position}{\code{IBREST2$get_position()}}
\item \href{#method-IBREST2-buy_and_confirm}{\code{IBREST2$buy_and_confirm()}}
\item \href{#method-IBREST2-cancel_order}{\code{IBREST2$cancel_order()}}
\item \href{#method-IBREST2-get_portfolio_summary}{\code{IBREST2$get_portfolio_summary()}}
\item \href{#method-IBREST2-clone}{\code{IBREST2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-new"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-new}{}}}
\subsection{Method \code{new()}}{
Create a new IBREST object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$new(domain = "localhost", port = 5000L, azure_storage_endpoint = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{domain}}{Domain, by default localhost}

\item{\code{port}}{Port, by default 5000}

\item{\code{azure_storage_endpoint}}{Azure storage endpoint}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `IBREST` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-ib_get"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-ib_get}{}}}
\subsection{Method \code{ib_get()}}{
IB GET object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$ib_get(
  url = modify_url(self$baseurl, path = "v1/api/sso/validate"),
  query = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{Url of GET endpoint.}

\item{\code{query}}{Query of GET request.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GET response.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-ib_post"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-ib_post}{}}}
\subsection{Method \code{ib_post()}}{
IB POST object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$ib_post(
  url = modify_url(self$baseurl, path = "v1/api/tickle"),
  body = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{Url of POST endpoint.}

\item{\code{body}}{Body of POST request.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GET response.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-get_api_spec"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-get_api_spec}{}}}
\subsection{Method \code{get_api_spec()}}{
Get IB REST API description.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$get_api_spec()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
JSON object with description.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-get_unadjusted_market"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-get_unadjusted_market}{}}}
\subsection{Method \code{get_unadjusted_market()}}{
Get unadjusted market data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$get_unadjusted_market(
  conid,
  exchange = NULL,
  period = "5d",
  bar = "1h",
  outsideRth = TRUE,
  keep_nytime_10_16 = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conid}}{Contract ID.}

\item{\code{exchange}}{Exchange.}

\item{\code{period}}{Period.}

\item{\code{bar}}{Bar.}

\item{\code{outsideRth}}{Data outside trading hours.}

\item{\code{keep_nytime_10_16}}{If TRUE, timezone is changed to NY time and only
trading hours kept.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with unadjusted market data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-get_position"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-get_position}{}}}
\subsection{Method \code{get_position()}}{
Get portfolio positions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$get_position(account_id, con_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_id}}{Account ID.}

\item{\code{con_id}}{Contract id.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list object with info on positions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-buy_and_confirm"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-buy_and_confirm}{}}}
\subsection{Method \code{buy_and_confirm()}}{
Place order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$buy_and_confirm(account_id, order_body)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_id}}{Account ID.}

\item{\code{order_body}}{Body of POST request which place order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list object with info on positions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-cancel_order"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-cancel_order}{}}}
\subsection{Method \code{cancel_order()}}{
Cancel order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$cancel_order(account_id, order_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_id}}{Account ID.}

\item{\code{order_id}}{Order id from Place orders endpoint result.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list object with info on positions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-get_portfolio_summary"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-get_portfolio_summary}{}}}
\subsection{Method \code{get_portfolio_summary()}}{
Portfolio summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$get_portfolio_summary(account_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_id}}{Account ID.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list object with info on podrtfolio summary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBREST2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IBREST2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBREST2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
