% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBTWS.R
\name{IBTWS}
\alias{IBTWS}
\title{IBTWS Class}
\description{
Get data data from IB.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{IBTWS}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{port}}{Port.}

\item{\code{clientId}}{Client ID.}

\item{\code{ic}}{Client ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IBTWS-new}{\code{IBTWS$new()}}
\item \href{#method-IBTWS-get_data}{\code{IBTWS$get_data()}}
\item \href{#method-IBTWS-clone}{\code{IBTWS$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBTWS-new"></a>}}
\if{latex}{\out{\hypertarget{method-IBTWS-new}{}}}
\subsection{Method \code{new()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBTWS$new(port = 4002, clientId = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{Port.}

\item{\code{clientId}}{client ID.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `IBTWS` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBTWS-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-IBTWS-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get market data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBTWS$get_data(days_history = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{days_history}}{how long to the past to look}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Earning announcements data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IBTWS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IBTWS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IBTWS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
