% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{Import}
\alias{Import}
\title{Import Class}
\description{
Help class that imports all relevant data from sources (for now onlz FMP).
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{Import}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fmp}}{Help field to hold FMP class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Import-new}{\code{Import$new()}}
\item \href{#method-Import-get_data_fmp}{\code{Import$get_data_fmp()}}
\item \href{#method-Import-clone}{\code{Import$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Import-new"></a>}}
\if{latex}{\out{\hypertarget{method-Import-new}{}}}
\subsection{Method \code{new()}}{
Create a new Import object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `Import` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Import-get_data_fmp"></a>}}
\if{latex}{\out{\hypertarget{method-Import-get_data_fmp}{}}}
\subsection{Method \code{get_data_fmp()}}{
Function imports all relevant data for USA stocks from FMP Prep.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$get_data_fmp(
  symbols,
  uri_market_cap = "F:/equity/usa/fundamentals/market_cap.parquet",
  uri_earning_announcements =
    "F:/data/equity/us/fundamentals/earning_announcements.parquet",
  uri_fundamentals = "F:/data/equity/us/fundamentals/fundamentals.parquet",
  uri_prices = "F:/data/equity/daily_fmp_all.csv",
  uri_dividends = "F:/equity/usa/fundamentals/dividends.parquet",
  first_date = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Securities symbols.}

\item{\code{uri_market_cap}}{Tiledb uri for market capitalization.}

\item{\code{uri_earning_announcements}}{Tiledb uri for earning announcements.}

\item{\code{uri_fundamentals}}{to fundamentals data.}

\item{\code{uri_prices}}{Tiledb uri for daily OHLCV prices.}

\item{\code{uri_dividends}}{Tiledb uri for dividends data.}

\item{\code{first_date}}{if NA, keep all. Otherwise, keep only data from
first_date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data.table with factors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Import-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Import-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Import$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
