% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lean.R
\name{Lean}
\alias{Lean}
\title{Lean Class}
\description{
Help functions for transforming raw market data to Quantconnect data.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{Lean}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Lean$new()}}
\item \href{#method-equity_hour_from_fmpcloud}{\code{Lean$equity_hour_from_fmpcloud()}}
\item \href{#method-equity_minute_from_fmpcloud}{\code{Lean$equity_minute_from_fmpcloud()}}
\item \href{#method-clone}{\code{Lean$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files">}\href{../../findata/html/DataAbstract.html#method-save_blob_files}{\code{findata::DataAbstract$save_blob_files()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Lean object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `Lean` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-equity_hour_from_fmpcloud"></a>}}
\if{latex}{\out{\hypertarget{method-equity_hour_from_fmpcloud}{}}}
\subsection{Method \code{equity_hour_from_fmpcloud()}}{
Convert FMP Cloud hour data to Quantconnect hour data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$equity_hour_from_fmpcloud(
  save_path = "D:/lean_projects/data/equity/usa/hour"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_path}}{Quantconnect data equity hour folder path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No valuie returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-equity_minute_from_fmpcloud"></a>}}
\if{latex}{\out{\hypertarget{method-equity_minute_from_fmpcloud}{}}}
\subsection{Method \code{equity_minute_from_fmpcloud()}}{
Convert FMP Cloud minute data to Quantconnect minute data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$equity_minute_from_fmpcloud(
  save_path = "D:/lean_projects/data/equity/usa/minute"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_path}}{Quantconnect data equity minute folder path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No value returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
