% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Macros.R
\name{Macros}
\alias{Macros}
\title{Macros Class}
\description{
Function calculates macro variables potentially important for
trading/investment strategies. Some are ussualy used in asset pricing
literature, while others are not.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{Macros}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{source}}{Data source for calculating macro predictors.}

\item{\code{fmp}}{Help field to hold FMP class.}

\item{\code{import}}{Help field to hold Import class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Macros-new}{\code{Macros$new()}}
\item \href{#method-Macros-get_macros}{\code{Macros$get_macros()}}
\item \href{#method-Macros-clone}{\code{Macros$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Macros-new"></a>}}
\if{latex}{\out{\hypertarget{method-Macros-new}{}}}
\subsection{Method \code{new()}}{
Create a new Macros object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Macros$new(source = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Data source for calculating macro predictors.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Macros` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Macros-get_macros"></a>}}
\if{latex}{\out{\hypertarget{method-Macros-get_macros}{}}}
\subsection{Method \code{get_macros()}}{
Main function that calculates factors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Macros$get_macros()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data.table with factors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Macros-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Macros-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Macros$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
