% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockAnalysis.R
\name{StockAnalysis}
\alias{StockAnalysis}
\title{StockAnalysis Class}
\description{
Get data data from StockAnalysis website.
}
\references{
\url{https://stockanalysis.com/actions/}
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{StockAnalysis}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_url}}{Base url}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StockAnalysis-new}{\code{StockAnalysis$new()}}
\item \href{#method-StockAnalysis-get_corporate_actions}{\code{StockAnalysis$get_corporate_actions()}}
\item \href{#method-StockAnalysis-get_corporate_tiledb}{\code{StockAnalysis$get_corporate_tiledb()}}
\item \href{#method-StockAnalysis-clone}{\code{StockAnalysis$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StockAnalysis-new"></a>}}
\if{latex}{\out{\hypertarget{method-StockAnalysis-new}{}}}
\subsection{Method \code{new()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StockAnalysis$new(
  base_url = "https://stockanalysis.com",
  azure_storage_endpoint = NULL,
  context_with_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{API KEY for FMP cloud data.}

\item{\code{azure_storage_endpoint}}{Azure storate endpont}

\item{\code{context_with_config}}{AWS S3 Tiledb config}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FMP` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StockAnalysis-get_corporate_actions"></a>}}
\if{latex}{\out{\hypertarget{method-StockAnalysis-get_corporate_actions}{}}}
\subsection{Method \code{get_corporate_actions()}}{
Get corporate action for symbol.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StockAnalysis$get_corporate_actions(
  k =
    "AAAAB3NzaC1yc2EAAAADAQABAAABAQDNK7jARJUOF5HofsvEkZi6T80FW4Shxx1k6tGnyw1bMyrGXOuMg7xx",
  query = c("listed", "delisted", "splits", "changes", "spinoffs", "bankruptcies",
    "acquisitions")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{Url id}

\item{\code{query}}{Type of corporate action}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table of corporate action.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StockAnalysis-get_corporate_tiledb"></a>}}
\if{latex}{\out{\hypertarget{method-StockAnalysis-get_corporate_tiledb}{}}}
\subsection{Method \code{get_corporate_tiledb()}}{
Get corporate actions and save it to TileDB..
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StockAnalysis$get_corporate_tiledb(url = "s3://equity-usa-corporateactions")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{TileDB uri}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Null
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StockAnalysis-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StockAnalysis-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StockAnalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
