% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tweeter.R
\name{Tweeter}
\alias{Tweeter}
\title{Tweeter Class}
\description{
Get data data from Tweeter.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_key}}{API KEY for FMP Cloud}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tweeter-new}{\code{Tweeter$new()}}
\item \href{#method-Tweeter-clone}{\code{Tweeter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tweeter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tweeter-new}{}}}
\subsection{Method \code{new()}}{
Create a new Tweeter object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tweeter$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `Tweeter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tweeter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Tweeter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tweeter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
