% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yields.R
\name{Yields}
\alias{Yields}
\title{Yields Class}
\description{
Get data on yields curve for various countries.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{countries}}{Countries to scrap data for. Default for all countres.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Yields-new}{\code{Yields$new()}}
\item \href{#method-Yields-get_yields}{\code{Yields$get_yields()}}
\item \href{#method-Yields-get_yields_us}{\code{Yields$get_yields_us()}}
\item \href{#method-Yields-get_yields_china}{\code{Yields$get_yields_china()}}
\item \href{#method-Yields-get_yields_eu}{\code{Yields$get_yields_eu()}}
\item \href{#method-Yields-clone}{\code{Yields$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Yields-new"></a>}}
\if{latex}{\out{\hypertarget{method-Yields-new}{}}}
\subsection{Method \code{new()}}{
Create a new Yields object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Yields$new(countries = c("US", "China", "EU"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{countries}}{character vector of countries to scrap data for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Yields` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Yields-get_yields"></a>}}
\if{latex}{\out{\hypertarget{method-Yields-get_yields}{}}}
\subsection{Method \code{get_yields()}}{
Scrap yields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Yields$get_yields(eu_yields_history_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eu_yields_history_file}}{yields histoty path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data.table with yields data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Yields-get_yields_us"></a>}}
\if{latex}{\out{\hypertarget{method-Yields-get_yields_us}{}}}
\subsection{Method \code{get_yields_us()}}{
Scrap yields for US.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Yields$get_yields_us()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data.table with yields data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Yields-get_yields_china"></a>}}
\if{latex}{\out{\hypertarget{method-Yields-get_yields_china}{}}}
\subsection{Method \code{get_yields_china()}}{
Scrap yields for China.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Yields$get_yields_china()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data.table with yields data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Yields-get_yields_eu"></a>}}
\if{latex}{\out{\hypertarget{method-Yields-get_yields_eu}{}}}
\subsection{Method \code{get_yields_eu()}}{
Scrap yields for EU from ECB.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Yields$get_yields_eu(yields_history_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{yields_history_file}}{yields histoty path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data.table with yields data for EU from ECB.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Yields-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Yields-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Yields$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
