% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finam.R
\name{Finam}
\alias{Finam}
\title{Finam Class}
\description{
Get data data from finam using Quanttools package.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{Finam}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Finam-new}{\code{Finam$new()}}
\item \href{#method-Finam-finam_tick_tiledb}{\code{Finam$finam_tick_tiledb()}}
\item \href{#method-Finam-clone}{\code{Finam$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Finam-new"></a>}}
\if{latex}{\out{\hypertarget{method-Finam-new}{}}}
\subsection{Method \code{new()}}{
Create a new Finam object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Finam$new(azure_storage_endpoint = NULL, context_with_config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{azure_storage_endpoint}}{Azure storate endpont}

\item{\code{context_with_config}}{AWS S3 Tiledb config}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Finam` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Finam-finam_tick_tiledb"></a>}}
\if{latex}{\out{\hypertarget{method-Finam-finam_tick_tiledb}{}}}
\subsection{Method \code{finam_tick_tiledb()}}{
Get tick data from finam source using QuantTools.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Finam$finam_tick_tiledb(
  symbols,
  days = getBusinessDays(as.Date("2011-01-01"), Sys.Date() - 1),
  save_uri = "s3://equity-usa-tick-finam"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Start date}

\item{\code{days}}{Days to scrap data for}

\item{\code{save_uri}}{Tiledb uri to save.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Get investing com ea data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Finam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Finam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Finam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
