% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finam.R
\name{Finam}
\alias{Finam}
\title{Finam Class}
\description{
Get data data from finam using Quanttools package.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path_to_dump}}{Local path to save data to.}

\item{\code{delay}}{Delay between requests.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Finam-new}{\code{Finam$new()}}
\item \href{#method-Finam-get_ticks}{\code{Finam$get_ticks()}}
\item \href{#method-Finam-clone}{\code{Finam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Finam-new"></a>}}
\if{latex}{\out{\hypertarget{method-Finam-new}{}}}
\subsection{Method \code{new()}}{
Create a new MacroData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Finam$new(path_to_dump, delay = 0.8)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_to_dump}}{Local path to save data to.}

\item{\code{delay}}{Delay between requests.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MacroData` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Finam-get_ticks"></a>}}
\if{latex}{\out{\hypertarget{method-Finam-get_ticks}{}}}
\subsection{Method \code{get_ticks()}}{
Get tick data from finam source using QuantTools.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Finam$get_ticks(
  symbol,
  days = getBusinessDays(as.Date("2010-01-01"), Sys.Date() - 1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Start date}

\item{\code{days}}{Days to scrap data for}

\item{\code{save_uri}}{Tiledb uri to save.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Get investing com ea data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Finam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Finam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Finam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
