% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finam.R
\name{Finam}
\alias{Finam}
\title{Finam Class}
\description{
Get data data from finam using Quanttools package.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{Finam}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Finam$new()}}
\item \href{#method-get_ticker_data}{\code{Finam$get_ticker_data()}}
\item \href{#method-clone}{\code{Finam$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files">}\href{../../findata/html/DataAbstract.html#method-save_blob_files}{\code{findata::DataAbstract$save_blob_files()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Finam object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Finam$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `Finam` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ticker_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_ticker_data}{}}}
\subsection{Method \code{get_ticker_data()}}{
Get tick data from finam source using QuantTools.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Finam$get_ticker_data(
  symbols,
  days = getBusinessDays(as.Date("2011-01-01"), Sys.Date() - 1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Start date}

\item{\code{days}}{Days to scrap data for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Get investing com ea data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Finam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
