% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMP.R
\name{FMP}
\alias{FMP}
\title{FMP Class}
\description{
Get data data from FMP Cloud.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{FMP}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_key}}{API KEY for FMP Cloud}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FMP-new}{\code{FMP$new()}}
\item \href{#method-FMP-get_earning_announcements}{\code{FMP$get_earning_announcements()}}
\item \href{#method-FMP-get_transcripts}{\code{FMP$get_transcripts()}}
\item \href{#method-FMP-get_daily_tiledb}{\code{FMP$get_daily_tiledb()}}
\item \href{#method-FMP-get_intraday_equities}{\code{FMP$get_intraday_equities()}}
\item \href{#method-FMP-create_start_end_dates}{\code{FMP$create_start_end_dates()}}
\item \href{#method-FMP-get_intraday_equities_batch}{\code{FMP$get_intraday_equities_batch()}}
\item \href{#method-FMP-get_sp500_symbols}{\code{FMP$get_sp500_symbols()}}
\item \href{#method-FMP-get_factor_files}{\code{FMP$get_factor_files()}}
\item \href{#method-FMP-get_ipo_date}{\code{FMP$get_ipo_date()}}
\item \href{#method-FMP-get_ipo_calendar_confirmed}{\code{FMP$get_ipo_calendar_confirmed()}}
\item \href{#method-FMP-get_fi_statement}{\code{FMP$get_fi_statement()}}
\item \href{#method-FMP-get_financial_metrics}{\code{FMP$get_financial_metrics()}}
\item \href{#method-FMP-get_stock_splits}{\code{FMP$get_stock_splits()}}
\item \href{#method-FMP-get_market_cap}{\code{FMP$get_market_cap()}}
\item \href{#method-FMP-get_market_cap_bulk}{\code{FMP$get_market_cap_bulk()}}
\item \href{#method-FMP-get_stock_list}{\code{FMP$get_stock_list()}}
\item \href{#method-FMP-get_symbol_changes}{\code{FMP$get_symbol_changes()}}
\item \href{#method-FMP-get_delisted_companies}{\code{FMP$get_delisted_companies()}}
\item \href{#method-FMP-get_available_traded_list}{\code{FMP$get_available_traded_list()}}
\item \href{#method-FMP-get_sp500_constituent}{\code{FMP$get_sp500_constituent()}}
\item \href{#method-FMP-get_fi_statement_bulk}{\code{FMP$get_fi_statement_bulk()}}
\item \href{#method-FMP-get_grades}{\code{FMP$get_grades()}}
\item \href{#method-FMP-get_targets}{\code{FMP$get_targets()}}
\item \href{#method-FMP-get_ratings}{\code{FMP$get_ratings()}}
\item \href{#method-FMP-get_historical_dividends}{\code{FMP$get_historical_dividends()}}
\item \href{#method-FMP-clone}{\code{FMP$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-new"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-new}{}}}
\subsection{Method \code{new()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$new(
  api_key = NULL,
  azure_storage_endpoint = NULL,
  context_with_config = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{API KEY for FMP cloud data.}

\item{\code{azure_storage_endpoint}}{Azure storate endpont}

\item{\code{context_with_config}}{AWS S3 Tiledb config}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FMP` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_earning_announcements"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_earning_announcements}{}}}
\subsection{Method \code{get_earning_announcements()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_earning_announcements(uri, start_date = NULL, consolidate = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{TileDB uri argument}

\item{\code{start_date}}{First date to scrape from.If NULL, takes last date from existing uri.}

\item{\code{consolidate}}{Consolidate and vacuum at the end.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Earning announcements data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_transcripts"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_transcripts}{}}}
\subsection{Method \code{get_transcripts()}}{
Get Earning Call Transcript from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_transcripts(
  symbols,
  years = 1992:2023,
  save_uri = "s3://equity-usa-transcripts"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{stock symbols. If tiledburi, import symbols from FMP events.}

\item{\code{years}}{years to extract data from.}

\item{\code{save_uri}}{Tiledb uri.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of GET request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_daily_tiledb"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_daily_tiledb}{}}}
\subsection{Method \code{get_daily_tiledb()}}{
get daily data from FMP cloud for all stocks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_daily_tiledb(uri = "s3://equity-usa-daily-fmp")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{AWS S3 url.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Scrap all daily data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_intraday_equities"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_intraday_equities}{}}}
\subsection{Method \code{get_intraday_equities()}}{
retrieving intraday market data from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_intraday_equities(
  symbol,
  multiply = 1,
  time = "day",
  from = as.character(Sys.Date() - 3),
  to = as.character(Sys.Date())
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol of the stock.}

\item{\code{multiply}}{multiplier.Start date.}

\item{\code{time}}{Size of the time. (minute, hour, day, week, month, quarter, year)}

\item{\code{from}}{Start date,}

\item{\code{to}}{End date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with ohlcv data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-create_start_end_dates"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-create_start_end_dates}{}}}
\subsection{Method \code{create_start_end_dates()}}{
Help function for calculating start and end dates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$create_start_end_dates(start_dates, n = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_dates}}{Sequence of start dates.}

\item{\code{n}}{number of consecutive dates.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of start and end dates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_intraday_equities_batch"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_intraday_equities_batch}{}}}
\subsection{Method \code{get_intraday_equities_batch()}}{
Get hour data for all history from fmp cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_intraday_equities_batch(
  symbols,
  url = "s3://equity-usa-minute-fmpcloud",
  save_uri_hour = "s3://equity-usa-hour-fmpcloud",
  save_uri_daily = "s3://equity-usa-daily-fmpcloud",
  deep_scan = FALSE,
  hardcode_start_dates = NULL,
  hardcode_end_dates = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Symbol of the stock.}

\item{\code{url}}{AWS S3 bucket url.}

\item{\code{save_uri_hour}}{AWS S3 bucket uri for hour data.}

\item{\code{save_uri_daily}}{AWS S3 bucket uri for daily data.}

\item{\code{deep_scan}}{should we test for dates with low number od observation
and try to scrap again.}

\item{\code{hardcode_start_dates}}{Start dates to scrape from.}

\item{\code{hardcode_end_dates}}{end_date to scrap from.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data saved to Azure blob.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_sp500_symbols"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_sp500_symbols}{}}}
\subsection{Method \code{get_sp500_symbols()}}{
Get SP500 symbols from fmp cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_sp500_symbols()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of SP500 symbols.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_factor_files"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_factor_files}{}}}
\subsection{Method \code{get_factor_files()}}{
Create factor files for calculating adjusted prices (adjusted for splits and dividends).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_factor_files(
  save_uri = "s3://equity-usa-factor-files",
  prices_uri = "s3://equity-usa-daily-fmpcloud"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_uri}}{Uri where tiledb saves files}

\item{\code{prices_uri}}{Uri to import prices from.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Factor files saved to "factor-file" blob.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_ipo_date"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_ipo_date}{}}}
\subsection{Method \code{get_ipo_date()}}{
Get IPO date from fmp cloud
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_ipo_date(ticker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticker}}{Stock ticker}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Stock IPO date.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_ipo_calendar_confirmed"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_ipo_calendar_confirmed}{}}}
\subsection{Method \code{get_ipo_calendar_confirmed()}}{
Get ipo-calendar-confirmed date from fmp cloud
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_ipo_calendar_confirmed(uri = "s3://equity-usa-ipo")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{TileDB uri argument}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Stock IPO date.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_fi_statement"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_fi_statement}{}}}
\subsection{Method \code{get_fi_statement()}}{
Get FI data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_fi_statement(
  symbol,
  statement = c("income-statement", "balance-sheet-statement", "cash-flow-statement"),
  period = c("annual", "quarter")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Stock symbol.}

\item{\code{statement}}{Stock statement.}

\item{\code{period}}{Stock period.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table of financial reports.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_financial_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_financial_metrics}{}}}
\subsection{Method \code{get_financial_metrics()}}{
Get financial ratios, company key metrics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_financial_metrics(
  symbol,
  type = c("ratios", "key-metrics", "financial-growth"),
  period = c("annual", "quarter")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Stock symbol.}

\item{\code{type}}{Type of fundamental analysis.}

\item{\code{period}}{Stock period.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table of financial reports.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_stock_splits"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_stock_splits}{}}}
\subsection{Method \code{get_stock_splits()}}{
Get stock splits data from FMP cloud Prep.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_stock_splits(start_date = Sys.Date() - 5, end_date = Sys.Date())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{Start date}

\item{\code{end_date}}{End date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table of stock splits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_market_cap"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_market_cap}{}}}
\subsection{Method \code{get_market_cap()}}{
Get market cap data from FMP cloud Prep.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_market_cap(ticker, limit = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticker}}{Ticker.}

\item{\code{limit}}{Limit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table with market cap data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_market_cap_bulk"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_market_cap_bulk}{}}}
\subsection{Method \code{get_market_cap_bulk()}}{
Get market cap data from FMP cloud Prep - bulk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_market_cap_bulk(uri = "s3://equity-usa-market-cap")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{TileDb uri.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table with market cap data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_stock_list"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_stock_list}{}}}
\subsection{Method \code{get_stock_list()}}{
Get stock list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_stock_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table with stock list data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_symbol_changes"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_symbol_changes}{}}}
\subsection{Method \code{get_symbol_changes()}}{
Get symbol changes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_symbol_changes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table with symbol changes data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_delisted_companies"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_delisted_companies}{}}}
\subsection{Method \code{get_delisted_companies()}}{
Get delisted companies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_delisted_companies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table with delisted companies data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_available_traded_list"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_available_traded_list}{}}}
\subsection{Method \code{get_available_traded_list()}}{
Get available traded list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_available_traded_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table with available traded list data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_sp500_constituent"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_sp500_constituent}{}}}
\subsection{Method \code{get_sp500_constituent()}}{
Get SP500 stocks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_sp500_constituent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table with SP500 tocks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_fi_statement_bulk"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_fi_statement_bulk}{}}}
\subsection{Method \code{get_fi_statement_bulk()}}{
Get Financial Statements data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_fi_statement_bulk(
  years = 1999:(data.table::year(Sys.Date())),
  statement = c("income-statement-bulk", "balance-sheet-statement-bulk",
    "cash-flow-statement-bulk", "ratios-bulk", "key-metrics-bulk",
    "financial-growth-bulk", "income-statement-growth-bulk",
    "balance-sheet-statement-growth-bulk", "cash-flow-statement-growth-bulk"),
  period = c("annual", "quarter")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{years}}{Get statement for specific year.}

\item{\code{statement}}{quarter or annual.}

\item{\code{period}}{quarter or annual.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table of financial reports.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_grades"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_grades}{}}}
\subsection{Method \code{get_grades()}}{
Get Upgrades and Downgrades from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_grades(symbols, save_uri = "s3://equity-usa-grades", update = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{stock symbols. If tiledburi, import symbols from FMP events.}

\item{\code{save_uri}}{Tiledb uri.}

\item{\code{update}}{Should we scrape all data od jut the new one.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of GET request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_targets"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_targets}{}}}
\subsection{Method \code{get_targets()}}{
Get Price Targest from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_targets(symbols, save_uri = "s3://equity-usa-targets", update = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{stock symbols. If tiledburi, import symbols from FMP events.}

\item{\code{save_uri}}{Tiledb uri.}

\item{\code{update}}{Should we scrape all data od jut the new one.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of GET request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_ratings"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_ratings}{}}}
\subsection{Method \code{get_ratings()}}{
Get Ratings from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_ratings(symbols, save_uri = "s3://equity-usa-ratings", update = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{stock symbols. If tiledburi, import symbols from FMP events.}

\item{\code{save_uri}}{Tiledb uri.}

\item{\code{update}}{Should we scrape all data od jut the new one.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of GET request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_historical_dividends"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_historical_dividends}{}}}
\subsection{Method \code{get_historical_dividends()}}{
Get Historical dividends from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_historical_dividends(symbols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Vector of stock symbols.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with historical dividends for symbol.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
