% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMP.R
\name{FMP}
\alias{FMP}
\title{FMP Class}
\description{
Get data data from FMP Cloud.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{FMP}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_key}}{API KEY for FMP Cloud}

\item{\code{azure_storage_endpoint}}{Azure storate endpont}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FMP-new}{\code{FMP$new()}}
\item \href{#method-FMP-get_earning_announcements}{\code{FMP$get_earning_announcements()}}
\item \href{#method-FMP-get_transcripts}{\code{FMP$get_transcripts()}}
\item \href{#method-FMP-update_transcripts}{\code{FMP$update_transcripts()}}
\item \href{#method-FMP-get_daily}{\code{FMP$get_daily()}}
\item \href{#method-FMP-get_daily_batch}{\code{FMP$get_daily_batch()}}
\item \href{#method-FMP-get_intraday_equities}{\code{FMP$get_intraday_equities()}}
\item \href{#method-FMP-get_intraday_equities_batch}{\code{FMP$get_intraday_equities_batch()}}
\item \href{#method-FMP-get_sp500_symbols}{\code{FMP$get_sp500_symbols()}}
\item \href{#method-FMP-get_factor_file}{\code{FMP$get_factor_file()}}
\item \href{#method-FMP-get_ipo_date}{\code{FMP$get_ipo_date()}}
\item \href{#method-FMP-get_fi_statement}{\code{FMP$get_fi_statement()}}
\item \href{#method-FMP-get_financial_metrics}{\code{FMP$get_financial_metrics()}}
\item \href{#method-FMP-get_stock_splits}{\code{FMP$get_stock_splits()}}
\item \href{#method-FMP-get_market_cap}{\code{FMP$get_market_cap()}}
\item \href{#method-FMP-get_market_cap_bulk}{\code{FMP$get_market_cap_bulk()}}
\item \href{#method-FMP-get_stock_list}{\code{FMP$get_stock_list()}}
\item \href{#method-FMP-get_available_traded_list}{\code{FMP$get_available_traded_list()}}
\item \href{#method-FMP-get_sp500_constituent}{\code{FMP$get_sp500_constituent()}}
\item \href{#method-FMP-clone}{\code{FMP$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-new"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-new}{}}}
\subsection{Method \code{new()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$new(api_key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{API KEY for FMP cloud data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FMP` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_earning_announcements"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_earning_announcements}{}}}
\subsection{Method \code{get_earning_announcements()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_earning_announcements(days_history = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{days_history}}{how long to the past to look}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Earning announcements data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_transcripts"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_transcripts}{}}}
\subsection{Method \code{get_transcripts()}}{
Get Earning Call Transcript from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_transcripts(
  symbols,
  years = 1992:2021,
  blob_cont = "transcripts",
  save_files = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{stock symbols}

\item{\code{years}}{years to extract data from.}

\item{\code{blob_cont}}{blob container}

\item{\code{save_files}}{should transcripts be saved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of GET request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-update_transcripts"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-update_transcripts}{}}}
\subsection{Method \code{update_transcripts()}}{
Update earnings announcements data from FMP cloud
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$update_transcripts(symbols, check_years = c(2021, 2022))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Sybmols}

\item{\code{check_years}}{Which years to check when updatingcheck_years}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return transcripts from FMP cloud.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_daily"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_daily}{}}}
\subsection{Method \code{get_daily()}}{
get daily data from FMP cloud
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_daily(symbols, date_start, date_end)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Symbols.}

\item{\code{date_start}}{Start date.}

\item{\code{date_end}}{End date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return transcripts from FMP cloud.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_daily_batch"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_daily_batch}{}}}
\subsection{Method \code{get_daily_batch()}}{
get daily data from FMP cloud for all stocks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_daily_batch()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Scrap all daily data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_intraday_equities"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_intraday_equities}{}}}
\subsection{Method \code{get_intraday_equities()}}{
retrieving intraday market data from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_intraday_equities(
  symbol,
  multiply = 1,
  time = "day",
  from = as.character(Sys.Date() - 3),
  to = as.character(Sys.Date())
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol of the stock.}

\item{\code{multiply}}{multiplier.Start date.}

\item{\code{time}}{Size of the time. (minute, hour, day, week, month, quarter, year)}

\item{\code{from}}{Start date,}

\item{\code{to}}{End date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with ohlcv data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_intraday_equities_batch"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_intraday_equities_batch}{}}}
\subsection{Method \code{get_intraday_equities_batch()}}{
Get hour data for all history from fmp cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_intraday_equities_batch(symbols, freq = c("hour", "minute"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Symbol of the stock.}

\item{\code{freq}}{frequency (hour or minute)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data saved to Azure blob.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_sp500_symbols"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_sp500_symbols}{}}}
\subsection{Method \code{get_sp500_symbols()}}{
Get SP500 symbols from fmp cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_sp500_symbols()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of SP500 symbols.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_factor_file"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_factor_file}{}}}
\subsection{Method \code{get_factor_file()}}{
Create factor files for calculating adjusted prices (adjusted for splits and dividends).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_factor_file()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Factor files saved to "factor-file" blob.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_ipo_date"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_ipo_date}{}}}
\subsection{Method \code{get_ipo_date()}}{
Get IPO date from fmp cloud
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_ipo_date(ticker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticker}}{Stock ticker}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Stock IPO date.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_fi_statement"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_fi_statement}{}}}
\subsection{Method \code{get_fi_statement()}}{
Get FI data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_fi_statement(
  symbol,
  statement = c("income-statement", "balance-sheet-statement", "cash-flow-statement"),
  period = c("annual", "quarter")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Stock symbol.}

\item{\code{statement}}{Stock statement.}

\item{\code{period}}{Stock period.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table of financial reports.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_financial_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_financial_metrics}{}}}
\subsection{Method \code{get_financial_metrics()}}{
Get financial ratios, company key metrics.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_financial_metrics(
  symbol,
  type = c("ratios", "key-metrics", "financial-growth"),
  period = c("annual", "quarter")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Stock symbol.}

\item{\code{type}}{Type of fundamental analysis.}

\item{\code{period}}{Stock period.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table of financial reports.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_stock_splits"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_stock_splits}{}}}
\subsection{Method \code{get_stock_splits()}}{
Get stock splits data from FMP cloud Prep.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_stock_splits(start_date = Sys.Date() - 5, end_date = Sys.Date())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{Start date}

\item{\code{end_date}}{End date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table of stock splits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_market_cap"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_market_cap}{}}}
\subsection{Method \code{get_market_cap()}}{
Get market cap data from FMP cloud Prep.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_market_cap(ticker, limit = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticker}}{Ticker.}

\item{\code{limit}}{Limit.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table with market cap data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_market_cap_bulk"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_market_cap_bulk}{}}}
\subsection{Method \code{get_market_cap_bulk()}}{
Get market cap data from FMP cloud Prep - bulk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_market_cap_bulk()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table with market cap data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_stock_list"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_stock_list}{}}}
\subsection{Method \code{get_stock_list()}}{
Get stock list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_stock_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table with stock list data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_available_traded_list"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_available_traded_list}{}}}
\subsection{Method \code{get_available_traded_list()}}{
Get available traded list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_available_traded_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table with available traded list data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-get_sp500_constituent"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-get_sp500_constituent}{}}}
\subsection{Method \code{get_sp500_constituent()}}{
Get SP500 stocks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_sp500_constituent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table with SP500 tocks.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FMP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FMP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
