% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMP.R
\name{FMP}
\alias{FMP}
\title{FMP Class}
\description{
Get data data from FMP Cloud.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{FMP}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_key}}{API KEY for FMP Cloud}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FMP$new()}}
\item \href{#method-get_earning_announcements}{\code{FMP$get_earning_announcements()}}
\item \href{#method-get_transcripts}{\code{FMP$get_transcripts()}}
\item \href{#method-update_transcripts}{\code{FMP$update_transcripts()}}
\item \href{#method-get_daily}{\code{FMP$get_daily()}}
\item \href{#method-get_daily_batch}{\code{FMP$get_daily_batch()}}
\item \href{#method-get_intraday_equities}{\code{FMP$get_intraday_equities()}}
\item \href{#method-get_intraday_equities_batch}{\code{FMP$get_intraday_equities_batch()}}
\item \href{#method-get_sp500_symbols}{\code{FMP$get_sp500_symbols()}}
\item \href{#method-clone}{\code{FMP$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files">}\href{../../findata/html/DataAbstract.html#method-save_blob_files}{\code{findata::DataAbstract$save_blob_files()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$new(api_key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{API KEY for FMP cloud data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FMP` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_earning_announcements"></a>}}
\if{latex}{\out{\hypertarget{method-get_earning_announcements}{}}}
\subsection{Method \code{get_earning_announcements()}}{
Create a new FMP object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_earning_announcements(days_history = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{days_history}}{how long to the past to look}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Earning announcements data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_transcripts"></a>}}
\if{latex}{\out{\hypertarget{method-get_transcripts}{}}}
\subsection{Method \code{get_transcripts()}}{
Get Earning Call Transcript from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_transcripts(
  symbols,
  years = 1992:2021,
  blob_cont = "transcripts",
  save_files = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{stock symbols}

\item{\code{years}}{years to extract data from.}

\item{\code{blob_cont}}{blob container}

\item{\code{save_files}}{should transcripts be saved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Result of GET request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_transcripts"></a>}}
\if{latex}{\out{\hypertarget{method-update_transcripts}{}}}
\subsection{Method \code{update_transcripts()}}{
Update earnings announcements data from FMP cloud
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$update_transcripts(symbols, check_years = c(2021, 2022))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Sybmols}

\item{\code{check_years}}{Which years to check when updatingcheck_years}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return transcripts from FMP cloud.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_daily"></a>}}
\if{latex}{\out{\hypertarget{method-get_daily}{}}}
\subsection{Method \code{get_daily()}}{
get daily data from FMP cloud
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_daily(symbols, date_start, date_end)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Symbols.}

\item{\code{date_start}}{Start date.}

\item{\code{date_end}}{End date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Return transcripts from FMP cloud.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_daily_batch"></a>}}
\if{latex}{\out{\hypertarget{method-get_daily_batch}{}}}
\subsection{Method \code{get_daily_batch()}}{
get daily data from FMP cloud for all stocks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_daily_batch()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Scrap all daily data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_intraday_equities"></a>}}
\if{latex}{\out{\hypertarget{method-get_intraday_equities}{}}}
\subsection{Method \code{get_intraday_equities()}}{
retrieving intraday market data from FMP cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_intraday_equities(
  symbol,
  multiply = 1,
  time = "minute",
  from = as.character(Sys.Date() - 3),
  to = as.character(Sys.Date())
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Symbol of the stock.}

\item{\code{multiply}}{multiplier.Start date.}

\item{\code{time}}{Size of the time. (minute, hour, day, week, month, quarter, year)}

\item{\code{from}}{Start date,}

\item{\code{to}}{End date.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with ohlcv data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_intraday_equities_batch"></a>}}
\if{latex}{\out{\hypertarget{method-get_intraday_equities_batch}{}}}
\subsection{Method \code{get_intraday_equities_batch()}}{
Get hour data for all history from fmp cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_intraday_equities_batch(symbols, freq = c("houe", "minute"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Symbol of the stock.}

\item{\code{freq}}{frequency (hour or minute)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data saved to Azure blob.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sp500_symbols"></a>}}
\if{latex}{\out{\hypertarget{method-get_sp500_symbols}{}}}
\subsection{Method \code{get_sp500_symbols()}}{
Get SP500 symbols from fmp cloud.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$get_sp500_symbols()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Vector of SP500 symbols.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FMP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
