% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lean.R
\name{Lean}
\alias{Lean}
\title{Lean Class}
\description{
Help functions for transforming raw market data to Quantconnect data.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{Lean}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lean-new}{\code{Lean$new()}}
\item \href{#method-Lean-equity_daily_from_fmpcloud}{\code{Lean$equity_daily_from_fmpcloud()}}
\item \href{#method-Lean-equity_hour_from_fmpcloud}{\code{Lean$equity_hour_from_fmpcloud()}}
\item \href{#method-Lean-equity_minute_from_fmpcloud}{\code{Lean$equity_minute_from_fmpcloud()}}
\item \href{#method-Lean-adjusted_qc_data}{\code{Lean$adjusted_qc_data()}}
\item \href{#method-Lean-clone}{\code{Lean$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lean-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lean-new}{}}}
\subsection{Method \code{new()}}{
Create a new Lean object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$new(azure_storage_endpoint = NULL, context_with_config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{azure_storage_endpoint}}{Azure storate endpont}

\item{\code{context_with_config}}{AWS S3 Tiledb config}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Lean` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lean-equity_daily_from_fmpcloud"></a>}}
\if{latex}{\out{\hypertarget{method-Lean-equity_daily_from_fmpcloud}{}}}
\subsection{Method \code{equity_daily_from_fmpcloud()}}{
Convert FMP Cloud daily data to Quantconnect daily data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$equity_daily_from_fmpcloud(
  lean_data_path = "D:/lean/data/equity/usa/daily",
  uri = "D:/equity-usa-daily-fmp",
  fast = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lean_data_path}}{Quantconnect data equity minute folder path.}

\item{\code{uri}}{TileDB uri argument.}

\item{\code{fast}}{If TRUE, all daily data is readed in RAM.
You should have at least 32 GB RAM to use this option for all US.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No value returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lean-equity_hour_from_fmpcloud"></a>}}
\if{latex}{\out{\hypertarget{method-Lean-equity_hour_from_fmpcloud}{}}}
\subsection{Method \code{equity_hour_from_fmpcloud()}}{
Convert FMP Cloud hour data to Quantconnect hour data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$equity_hour_from_fmpcloud(
  save_path = "D:/lean_projects/data/equity/usa/hour"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_path}}{Quantconnect data equity hour folder path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No valuie returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lean-equity_minute_from_fmpcloud"></a>}}
\if{latex}{\out{\hypertarget{method-Lean-equity_minute_from_fmpcloud}{}}}
\subsection{Method \code{equity_minute_from_fmpcloud()}}{
Convert FMP Cloud minute data to Quantconnect minute data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$equity_minute_from_fmpcloud(
  lean_data_path = "D:/lean_projects/data/equity/usa/minute",
  uri = "D:/equity-usa-minute-fmpcloud",
  uri_factor_files = "s3://equity-usa-factor-files"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lean_data_path}}{Quantconnect data equity minute folder path.}

\item{\code{uri}}{TileDB uri argument.}

\item{\code{uri_factor_files}}{TileDB uri argument for factor files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No value returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lean-adjusted_qc_data"></a>}}
\if{latex}{\out{\hypertarget{method-Lean-adjusted_qc_data}{}}}
\subsection{Method \code{adjusted_qc_data()}}{
Adjust bough QuantConnect. Use map_files and factor_files
    to adjust OHLCV data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$adjusted_qc_data(path_market_data, path_map_files, path_factor_files)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_market_data}}{Path to daily / hourly data.}

\item{\code{path_map_files}}{Path to map files.}

\item{\code{path_factor_files}}{Path to factor files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with adjusted QC data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lean-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Lean-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
