% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsData.R
\name{UtilsData}
\alias{UtilsData}
\title{UtilsData Class}
\description{
Various help data.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{UtilsData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UtilsData$new()}}
\item \href{#method-get_ticker_data}{\code{UtilsData$get_ticker_data()}}
\item \href{#method-clone}{\code{UtilsData$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files">}\href{../../findata/html/DataAbstract.html#method-save_blob_files}{\code{findata::DataAbstract$save_blob_files()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new UtilsData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `UtilsData` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ticker_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_ticker_data}{}}}
\subsection{Method \code{get_ticker_data()}}{
Get tick data from finam source using QuantTools.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$get_ticker_data(symbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Equity symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Get investing com ea data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
