% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsData.R
\name{UtilsData}
\alias{UtilsData}
\title{UtilsData Class}
\description{
Various help data.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{UtilsData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtilsData-new}{\code{UtilsData$new()}}
\item \href{#method-UtilsData-get_ticker_data}{\code{UtilsData$get_ticker_data()}}
\item \href{#method-UtilsData-get_split_factor}{\code{UtilsData$get_split_factor()}}
\item \href{#method-UtilsData-adjust_fm_tiledb}{\code{UtilsData$adjust_fm_tiledb()}}
\item \href{#method-UtilsData-sp500_history}{\code{UtilsData$sp500_history()}}
\item \href{#method-UtilsData-get_market_data}{\code{UtilsData$get_market_data()}}
\item \href{#method-UtilsData-adjust_market_data}{\code{UtilsData$adjust_market_data()}}
\item \href{#method-UtilsData-get_map_files_tiledb}{\code{UtilsData$get_map_files_tiledb()}}
\item \href{#method-UtilsData-clone}{\code{UtilsData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-new}{}}}
\subsection{Method \code{new()}}{
Create a new UtilsData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$new(azure_storage_endpoint = NULL, context_with_config = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{azure_storage_endpoint}}{Azure storate endpont}

\item{\code{context_with_config}}{AWS S3 Tiledb config}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `UtilsData` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-get_ticker_data"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-get_ticker_data}{}}}
\subsection{Method \code{get_ticker_data()}}{
Get tick data from finam source using QuantTools.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$get_ticker_data(symbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Equity symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Get investing com ea data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-get_split_factor"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-get_split_factor}{}}}
\subsection{Method \code{get_split_factor()}}{
Get split data from the web and calculate split factors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$get_split_factor(ticker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticker}}{Stock ticker}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Stock splits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-adjust_fm_tiledb"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-adjust_fm_tiledb}{}}}
\subsection{Method \code{adjust_fm_tiledb()}}{
Adjusted instraday data for splits and dividends
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$adjust_fm_tiledb(
  save_uri = "D:/equity-usa-minute-fmpcloud-adjusted",
  minute_uri = "D:/equity-usa-minute-fmpcloud"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_uri}}{TileDB uri for saving}

\item{\code{minute_uri}}{TileDB uri with unadjusted minute data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Adjusted data saved to AWS S3 bucket
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-sp500_history"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-sp500_history}{}}}
\subsection{Method \code{sp500_history()}}{
Get SP500 stocks for every date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$sp500_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data table of Sp500 tickers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-get_market_data"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-get_market_data}{}}}
\subsection{Method \code{get_market_data()}}{
Help function to import clean data from Azureblob.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$get_market_data(symbols = c("aapl", "aal"), frequency = "hour")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Symbols}

\item{\code{frequency}}{Frequency can be hour and minute.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with market prices
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-adjust_market_data"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-adjust_market_data}{}}}
\subsection{Method \code{adjust_market_data()}}{
adjust market data for splits and/or dividends
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$adjust_market_data(
  market_data,
  factor_files,
  type = c("all", "dividends", "splits"),
  cols_adjust = c("o", "h", "l", "c")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{market_data}}{Market data of OHLCV type.}

\item{\code{factor_files}}{Factor files must contain columns:
- symbol
- date
- price_factor
- split_factor}

\item{\code{type}}{Type of adjustment.}

\item{\code{cols_adjust}}{Columns to adjust.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with adjusted market prices.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-get_map_files_tiledb"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-get_map_files_tiledb}{}}}
\subsection{Method \code{get_map_files_tiledb()}}{
Help function to get map files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$get_map_files_tiledb(
  save_uri = "s3://equity-usa-mapfiles",
  symbol_uri = "https://en.wikipedia.org/wiki/S\%26P_100"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_uri}}{TileDB uri.}

\item{\code{symbol_uri}}{TileDB uri for saved files}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Maped files.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
