% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsData.R
\name{UtilsData}
\alias{UtilsData}
\title{UtilsData Class}
\description{
Various help data.
}
\section{Super class}{
\code{\link[findata:DataAbstract]{findata::DataAbstract}} -> \code{UtilsData}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{azure_storage_endpoint}}{Azure storate endpont}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtilsData-new}{\code{UtilsData$new()}}
\item \href{#method-UtilsData-get_ticker_data}{\code{UtilsData$get_ticker_data()}}
\item \href{#method-UtilsData-get_split_factor}{\code{UtilsData$get_split_factor()}}
\item \href{#method-UtilsData-adjust_intraday_data}{\code{UtilsData$adjust_intraday_data()}}
\item \href{#method-UtilsData-sp500_history}{\code{UtilsData$sp500_history()}}
\item \href{#method-UtilsData-get_market_data}{\code{UtilsData$get_market_data()}}
\item \href{#method-UtilsData-clone}{\code{UtilsData$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="findata" data-topic="DataAbstract" data-id="save_blob_files"><a href='../../findata/html/DataAbstract.html#method-DataAbstract-save_blob_files'><code>findata::DataAbstract$save_blob_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-new}{}}}
\subsection{Method \code{new()}}{
Create a new UtilsData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new `UtilsData` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-get_ticker_data"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-get_ticker_data}{}}}
\subsection{Method \code{get_ticker_data()}}{
Get tick data from finam source using QuantTools.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$get_ticker_data(symbol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{Equity symbol}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Get investing com ea data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-get_split_factor"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-get_split_factor}{}}}
\subsection{Method \code{get_split_factor()}}{
Get split data from the web and calculate split factors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$get_split_factor(ticker)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticker}}{Stock ticker}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Stock splits
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-adjust_intraday_data"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-adjust_intraday_data}{}}}
\subsection{Method \code{adjust_intraday_data()}}{
Adjusted instraday data for splits and dividends
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$adjust_intraday_data(freq = c("hour", "minute"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{freq}}{Frequency, can be hour or minute.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Adjusted data saved to Azure blob
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-sp500_history"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-sp500_history}{}}}
\subsection{Method \code{sp500_history()}}{
Get SP500 stocks for every date.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$sp500_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data table of Sp500 tickers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-get_market_data"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-get_market_data}{}}}
\subsection{Method \code{get_market_data()}}{
Help function to import clean data from Azureblob.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$get_market_data(symbols = c("aapl", "aal"), frequency = "hour")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbols}}{Symbols}

\item{\code{frequency}}{Frequency can be hour and minute.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with market prices
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtilsData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UtilsData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtilsData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
